adding failing tests
