+++ test/utils.js
@@ -13,10 +13,10 @@ const assert = require('assert');
 const path = require('path');
 const shell = require('shelljs');
 const fse = require('fs-extra');
-const unzip = require('unzip');
 const http = require('http');
 const Replay = require('replay');
 const uuidv4 = require('uuid/v4');
+const unzip = require('unzip2');
 const winston = require('winston');
 const BuildCommand = require('../src/build.cmd');
 
@@ -100,6 +100,28 @@ async function assertHttp(url, status, spec, replacements = []) {
     }).on('error', (e) => {
       reject(e);
     });
+  });
+}
+
+async function assertZipEntries(zipPath, entries) {
+  assertFile(zipPath);
+
+  // check zip
+  const result = await new Promise((resolve, reject) => {
+    const es = [];
+    const srcStream = fse.createReadStream(zipPath);
+    srcStream.pipe(unzip.Parse())
+      .on('entry', (entry) => {
+        es.push(entry.path);
+        entry.autodrain();
+      })
+      .on('close', () => {
+        resolve(es);
+      })
+      .on('error', reject);
+  });
+  entries.forEach((s) => {
+    assert.ok(result.indexOf(s) >= 0, `${s} must be included in ${zipPath}`);
   });
 }
 
@@ -272,4 +294,5 @@ module.exports = {
   createLogger,
   processSource,
   perfExample,
+  assertZipEntries,
 };
